fix(compile): 修复 7 类编译错误，首次真实编译通过
GitHub Actions（Xcode 16.4 / iPhoneOS18.5.sdk / arm64-apple-ios17.0）
首次真实编译暴露的问题，均为既有代码、此前从未被编译过：

1. BLEProtocol.swift:14-20  cannot find 'CBUUID' in scope
   → 补 import CoreBluetooth
2. BLEProtocol.swift:39  type 'UInt16' has no member 'addReportingOverflow'
   → 改为实例方法 s.addReportingOverflow(UInt16(b))，校验和语义不变
3. PatternRenderer.swift:16,25  UIColor / RendererContext 没有 fill
   → 改用 ctx.cgContext (+ setFillColor(cgColor)) 绘制
4. PixelConverter.swift:30-46  Int/CGFloat 混用
   → side 提升为 CGFloat，gw/gh 显式 Int()/CGFloat() 转换
5. BoardTabView.swift:263  Section 重载歧义致泛型推断失败
   → 字符串标题移入 header: 闭包
6. ConvertView.swift:231  UIImage? 未解包
   → 加 guard let image else { return }（不强解包）
7. FolderViews.swift:249,560,572  对非可选 Result 用 if let/guard let
   → 改判 res.cells.isEmpty，保持原成败分支行为

警告修复：
- BLECentral.swift boardNameHints 标 nonisolated（Swift 6 下将成错误）
- BoardSession.swift 未使用的 ack 绑定改为 != nil 判断

+++ Sources/Core/PixelConverter.swift
@@ -27,13 +27,13 @@ enum PixelConverter {
 
         // 按长宽比计算网格尺寸
         let srcW = CGFloat(cg.width), srcH = CGFloat(cg.height)
-        let side = max(8, min(104, options.maxSide))
+        let side = CGFloat(max(8, min(104, options.maxSide)))
         var gw: Int, gh: Int
         if srcW >= srcH {
-            gw = side
+            gw = Int(side)
             gh = max(8, Int(round(side * srcH / srcW)))
         } else {
-            gh = side
+            gh = Int(side)
             gw = max(8, Int(round(side * srcW / srcH)))
         }
 
@@ -43,7 +43,7 @@ enum PixelConverter {
             return Result(width: 0, height: 0, cells: [])
         }
         ctx.interpolationQuality = .medium
-        ctx.draw(cg, in: CGRect(x: 0, y: 0, width: gw, height: gh))
+        ctx.draw(cg, in: CGRect(x: 0, y: 0, width: CGFloat(gw), height: CGFloat(gh)))
         guard let data = ctx.data else { return Result(width: 0, height: 0, cells: []) }
 
         let buf = data.bindMemory(to: UInt8.self, capacity: gw * gh * 4)
